+++ src/babashka/impl/classes.clj
@@ -625,9 +625,7 @@
           org.jsoup.nodes.TextNode
           org.jsoup.nodes.XmlDeclaration
           org.jsoup.parser.Tag
-          org.jsoup.parser.Parser
-          com.fasterxml.jackson.core.JsonGenerator
-          com.fasterxml.jackson.core.JsonParseException]
+          org.jsoup.parser.Parser]
     :constructors [clojure.lang.Delay
                    clojure.lang.DynamicClassLoader
                    clojure.lang.LineNumberingPushbackReader